Temporary move into the esp32/ directory
Later can be moved to a common location
