+++ src/lib/core/CoreLayer.am
@@ -32,7 +32,6 @@ CHIP_BUILD_CORE_LAYER_SOURCE_FILES                        = \
     @top_builddir@/src/lib/core/CHIPTLVWriter.cpp           \
     @top_builddir@/src/lib/core/CHIPTLVUpdater.cpp          \
     @top_builddir@/src/lib/core/CHIPKeyIds.cpp              \
-    @top_builddir@/src/lib/core/CHIPDeviceManager.cpp       \
     $(NULL)
 
 CHIP_BUILD_CORE_LAYER_HEADER_FILES                        = \
@@ -55,5 +54,4 @@ CHIP_BUILD_CORE_LAYER_HEADER_FILES                        = \
     @top_builddir@/src/lib/core/Optional.h                  \
     @top_builddir@/src/lib/core/ReferenceCounted.h          \
     @top_builddir@/src/lib/core/CHIPVendorIdentifiers.hpp   \
-    @top_builddir@/src/lib/core/CHIPDeviceManager.h         \
     $(NULL)
